refactor: Update test suite structure and consolidate test files for PlatformIO compatibility
